Small refactor before EditPersonalInfo page - User types refactor
