+++ singular-web/src/app/app.routes.ts
@@ -24,7 +24,7 @@ export const routes: Routes = [
     loadComponent: () => import('./layout/root/root.component').then((m) => m.RootComponent),
     canActivate: [authGuard],
     children: [
-      { path: '', pathMatch: 'full', redirectTo: 'profile' },
+      { path: '', pathMatch: 'full', redirectTo: 'home' },
       {
         path: 'home',
         loadComponent: () => import('./pages/home/home.compontent').then((m) => m.HomeComponent),
