+++ singular-web/src/app/auth/state/auth.reducer.ts
@@ -1,9 +1,9 @@
 import { createReducer, on } from '@ngrx/store';
 import { AuthActions } from './auth.actions';
-import { UserResponse } from '../../core/types';
+import { User } from '../../core/types';
 
 export interface AuthState {
-  user: UserResponse | null;
+  user: User | null;
   loading: boolean;
   error: string | null;
 }
