+++ singular-web/src/app/core/types/index.ts
@@ -1,1 +1,2 @@
 export * from './auth.types';
+export * from './user.types';
